refactor: 🔨이메일 전송 컨트롤러 이름 수정
